double for s

+++ SQLite.cs
@@ -145,7 +145,7 @@ namespace Palantir
     {
         [Key]
         public int Login { get; set; }
-        public int StartUTCS { get; set; }
+        public double StartUTCS { get; set; }
         public int DurationS { get; set; }
         public double Factor { get; set; }
     }
